refactor(parser): Polish the DiameterMessageParser
I value JavaDoc and a clean API and less repetition. This thing is
intended to be used with Netty in the end so a ByteBuf API seems the
most fitting.

You can now parse an entire message while also getting an idea on how
much information to parse in the first place. None of this modifies the
actual bytebuffer position so that's the responsibility of the outside
class.