feat(req/ans): Use factory methods + flag defaults
There are a lot of flags on a Diameter message and instead of putting
the burden on the user to set them correctly, the lib should take care
of this. E.g. the retransmission flag does not make sense on answers and
is disallowed. Also CER / CEA and DWR / DWA must never be proxiable.

There are standards compliant defaults now and the remaining flags can
be controlled with clearly marked factory methods.