feat(avp): Add new AVP infrastructure
This allows for creation of AVPs in a standards compliant way that is
also extensible. There are still things TODO though. Mainly we need to
implement all AVP value types that are missing and also update the
CodeAVPProvider. Also documentation for how to extend the AVPFactory
through own providers may be helpful.

Oh and: We need to use the Factory in the relevant setters of our own
requests / answers.