feat: Implement remaining RFC  diameter messages
All Requests / Answers of RFC 6733 have been added here. This also
includes a major re-factoring as I have decided that it would be more
elegant to have the defined AVPs as Mixins. That way, shared AVPs have
the same signature and behavior in any type of command.

Their API is also generic and allows for fluent setters.